--- a/deno.jsonc
+++ b/deno.jsonc
@@ -25,6 +25,7 @@
       "LICENSE"
     ],
     "exclude": [
+      "**/*_bench.ts",
       "**/*_test.ts",
       ".*"
     ]