--- a/mutex_bench.ts
+++ b/mutex_bench.ts
@@ -1,0 +1,7 @@
+import { Mutex } from "./mutex.ts";
+
+const mutex = new Mutex();
+
+Deno.bench("Mutex", async function () {
+  using _lock = await mutex.acquire();
+});