--- a/semaphore_bench.ts
+++ b/semaphore_bench.ts
@@ -1,0 +1,12 @@
+import { Semaphore } from "./semaphore.ts";
+
+const sem = new Semaphore(5);
+
+Deno.bench("Semaphore", async function () {
+  const worker = () => {
+    return sem.lock(async () => {
+      // do nothing
+    });
+  };
+  await Promise.all(Array.from({ length: 100 }).map(() => worker()));
+});