Performance improvement refactoring suggested in https://github.com/jsr-core/asyncutil/issues/30 are applied.

```
Check file:///Users/alisue/ogh/jsr-core/asyncutil/mutex_bench.ts
Check file:///Users/alisue/ogh/jsr-core/asyncutil/semaphore_bench.ts
cpu: Apple M1 Max
runtime: deno 1.45.4 (aarch64-apple-darwin)

file:///Users/alisue/ogh/jsr-core/asyncutil/mutex_bench.ts
benchmark              time (avg)        iter/s             (min … max)       p75       p99      p995
----------------------------------------------------------------------- -----------------------------
Mutex                 452.41 ns/iter   2,210,370.0 (439.27 ns … 545.97 ns) 446.38 ns 537.83 ns 545.97 ns
Mutex (1.0.2)         666.79 ns/iter   1,499,722.7 (636.54 ns … 947.56 ns) 653.95 ns 947.56 ns 947.56 ns
Mutex (Issue #30)     435.85 ns/iter   2,294,381.9  (428.7 ns … 458.39 ns) 435.39 ns 457.76 ns 458.39 ns


file:///Users/alisue/ogh/jsr-core/asyncutil/semaphore_bench.ts
benchmark              time (avg)        iter/s             (min … max)       p75       p99      p995
----------------------------------------------------------------------- -----------------------------
Semaphore              78.77 µs/iter      12,695.7    (61.62 µs … 1.81 ms) 68.04 µs 607.38 µs 665.96 µs
Semaphore (1.0.2)     296.69 µs/iter       3,370.6    (260.5 µs … 2.24 ms) 279.54 µs 635.88 µs 687.54 µs
```

Thanks to @PandaWorker 🎉