@PandaWorker 2024-08-15
Also, I would recommend using under the hood of the module only: lock\[Symbol.dispose\]()

And give customers the opportunity to manually release

```ts
const sem = new Semaphore(1);

{
  const lock = await sem.acquire();
  try {
     // do ...
  } finally {
    lock[Symbol.dispose](); // ot sem.release();
  }
}

{
  const lock = await sem.acquire();
  return asyncTask().finally(() => lock[Symbol.dispose]); // or sem.release();
}
```

You also need to write all possible tests and benchmarks