@PandaWorker 2024-08-15
Also Denon.bench does not run tasks competitively, which means that the semaphore will always be unlocked and will not put the expectant in the Set

You also need to make benchmarks for competitive access to acquire, so that the expectations are exactly put in the Set