@PandaWorker 2024-08-15
The semaphore implementation on Set will work in much the same way as on Deque, as long as there are few elements. As soon as there are a lot of expectations, there is a significant performance drawdown, since Set hashes the element itself when adding and searching for an element, and Deque immediately gives the first element

```
cpu: Apple M1 Max
runtime: deno 1.45.5 (aarch64-apple-darwin)

file:///Users/panda/Documents/deno/sync/semaphore.bench.ts
benchmark                                                      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------------------------------------------------------- -----------------------------

group semaphore acquire
relesae with using                                            631.12 ns/iter   1,584,480.8 (456.72 ns … 881.66 ns) 657.79 ns 881.66 ns 881.66 ns
relesae with using [acquireWithSignal]                        584.76 ns/iter   1,710,112.1 (474.36 ns … 929.88 ns) 627.16 ns 929.88 ns 929.88 ns
relesae with [Symbol.dispose]                                  373.9 ns/iter   2,674,518.7 (292.75 ns … 459.38 ns) 421.79 ns 457.5 ns 459.38 ns
relesae with [Symbol.dispose] (Deque)                         417.88 ns/iter   2,393,042.3  (320.25 ns … 512.5 ns) 466.82 ns 508.17 ns 512.5 ns
relesae with [Symbol.dispose] [acquireWithSignal]                422 ns/iter   2,369,692.6   (294.93 ns … 2.37 µs) 424.22 ns 1.82 µs 2.37 µs
relesae with [Symbol.dispose] [acquireWithSignal] (Deque)     409.78 ns/iter   2,440,341.3   (302.27 ns … 2.41 µs) 419.23 ns 1.51 µs 2.41 µs

summary
  relesae with [Symbol.dispose]
   1.1x faster than relesae with [Symbol.dispose] [acquireWithSignal] (Deque)
   1.12x faster than relesae with [Symbol.dispose] (Deque)
   1.13x faster than relesae with [Symbol.dispose] [acquireWithSignal]
   1.56x faster than relesae with using [acquireWithSignal]
   1.69x faster than relesae with using

group semaphore concurents tasks (5)
relesae with using                                              3.84 µs/iter     260,238.5     (3.23 µs … 4.71 µs) 4.07 µs 4.71 µs 4.71 µs
relesae with using [acquireWithSignal]                          3.91 µs/iter     255,575.8     (3.01 µs … 5.06 µs) 4.18 µs 5.06 µs 5.06 µs
relesae with [Symbol.dispose]                                   2.82 µs/iter     354,830.3     (2.31 µs … 3.17 µs) 2.92 µs 3.17 µs 3.17 µs
relesae with [Symbol.dispose] (Deque)                           2.49 µs/iter     402,009.5      (1.9 µs … 2.71 µs) 2.66 µs 2.71 µs 2.71 µs
relesae with [Symbol.dispose] [acquireWithSignal]               2.72 µs/iter     368,041.4     (2.22 µs … 3.41 µs) 3.01 µs 3.41 µs 3.41 µs
relesae with [Symbol.dispose] [acquireWithSignal] (Deque)       2.53 µs/iter     395,288.0     (2.01 µs … 2.91 µs) 2.71 µs 2.91 µs 2.91 µs

summary
  relesae with [Symbol.dispose] (Deque)
   1.02x faster than relesae with [Symbol.dispose] [acquireWithSignal] (Deque)
   1.09x faster than relesae with [Symbol.dispose] [acquireWithSignal]
   1.13x faster than relesae with [Symbol.dispose]
   1.54x faster than relesae with using
   1.57x faster than relesae with using [acquireWithSignal]

group semaphore concurents tasks (10_000)
relesae with using                                             34.14 ms/iter          29.3  (25.57 ms … 112.56 ms) 32.61 ms 112.56 ms 112.56 ms
relesae with using [acquireWithSignal]                         44.94 ms/iter          22.3    (26.4 ms … 119.2 ms) 56.11 ms 119.2 ms 119.2 ms
relesae with [Symbol.dispose]                                  27.16 ms/iter          36.8   (21.55 ms … 58.43 ms) 28.34 ms 58.43 ms 58.43 ms
relesae with [Symbol.dispose] (Deque)                            6.2 ms/iter         161.3    (4.62 ms … 15.66 ms) 6.61 ms 15.66 ms 15.66 ms
relesae with [Symbol.dispose] [acquireWithSignal]              25.95 ms/iter          38.5   (21.04 ms … 34.29 ms) 28.16 ms 34.29 ms 34.29 ms
relesae with [Symbol.dispose] [acquireWithSignal] (Deque)        5.8 ms/iter         172.4    (3.65 ms … 18.56 ms) 6.66 ms 18.56 ms 18.56 ms

summary
  relesae with [Symbol.dispose] [acquireWithSignal] (Deque)
   1.07x faster than relesae with [Symbol.dispose] (Deque)
   4.47x faster than relesae with [Symbol.dispose] [acquireWithSignal]
   4.68x faster than relesae with [Symbol.dispose]
   5.88x faster than relesae with using
   7.75x faster than relesae with using [acquireWithSignal]

group semaphore concurents tasks (50_000)
relesae with using                                            687.02 ms/iter           1.5  (650.5 ms … 739.51 ms) 695.77 ms 739.51 ms 739.51 ms
relesae with using [acquireWithSignal]                        702.04 ms/iter           1.4 (591.96 ms … 835.46 ms) 787.13 ms 835.46 ms 835.46 ms
relesae with [Symbol.dispose]                                 610.42 ms/iter           1.6  (574.81 ms … 643.3 ms) 629.11 ms 643.3 ms 643.3 ms
relesae with [Symbol.dispose] (Deque)                          66.74 ms/iter          15.0    (57.17 ms … 81.5 ms) 68.79 ms 81.5 ms 81.5 ms
relesae with [Symbol.dispose] [acquireWithSignal]             628.18 ms/iter           1.6 (549.56 ms … 676.17 ms) 671.43 ms 676.17 ms 676.17 ms
relesae with [Symbol.dispose] [acquireWithSignal] (Deque)      85.02 ms/iter          11.8  (65.58 ms … 105.92 ms) 92.25 ms 105.92 ms 105.92 ms

summary
  relesae with [Symbol.dispose] (Deque)
   1.27x faster than relesae with [Symbol.dispose] [acquireWithSignal] (Deque)
   9.15x faster than relesae with [Symbol.dispose]
   9.41x faster than relesae with [Symbol.dispose] [acquireWithSignal]
   10.29x faster than relesae with using
   10.52x faster than relesae with using [acquireWithSignal]
```