@lambdalisue 2024-08-15
> It's not entirely clear what Object.assign is used for. When I measured performance, the version with Object.assign was slower because additional instructions were executed, which in principle are not needed at all
> 
> Explain to me why to bind the disposing methods to the promise itself?
> 
> ```ts
> return Object.assign(Promise.resolve(disposable), disposable);
> return Object.assign(waiter.promise.then(() => disposable), disposable);
> ```

The current `RwLock` implementation relies on that.