@lambdalisue 2024-08-15
> While there is no native using feature in v8, you should not use it in places where it can harm performance and in libraries

I see. I haven't considered that. Now, I think we need to add `release()` for convenience